se mueve logica de llamado a la data al componente padre y se entrega como props array con pokemones
