+++ Pokedex/src/pages/home.vue
@@ -5,19 +5,34 @@
 </nav>
   <div>
     <formSearch />
-    <pokemonList />
+    <pokemonList  :pokemones="pokemones"/>
   </div>
 </template>
 
 <script>
+
+import { ref } from 'vue';
+import { fetchPokemonData } from '../components/api';
 import formSearch from '../components/fromSearch.vue';
 import pokemonList from '../components/pokemon.vue';
+
+const pokemones = ref(null);
 
 export default {
   components: {
     formSearch,
     pokemonList,
   },
- 
-};
+  data() {
+    return {
+      pokemones: [],
+    };
+  },
+   async created() {
+    pokemones.value = await fetchPokemonData();
+    console.log(pokemones.value.results, "setup")
+    this.pokemones = pokemones.value
+    // return pokemones
+  }
+}
 </script>
